Fix Bodega Client plugin system
Right now the bodega-client's handler plugin loading was broken.
Every Asset needs to have its own AssetHandler, but we were
requesting the instances to the QPluginLoader who would return us
every time the same handler instance.

This patch changes this by making the plugin to offer a factory.
This factory will be in charge of creating the handle instance.
It also adapts the handlers to this new behavior.

REVIEW: 107407